add few more stores
