--- Onbording/onboarding_items.dart
@@ -1,24 +1,0 @@
-import 'package:clothing_app/Onbording/onbording_info.dart';
-
-class OnboardingItems {
-  List<OnboardingInfo> items = [
-    OnboardingInfo(
-      title: "Welcome to Dress Me",
-      description:
-          "Our app is designed to help you create clothing that matches your unique style and preferences. Simply answer a few quick questions about your occasion, fabric choice, and design ideas, and let us guide you through the process.",
-      image: "assets/images/image1.jpg",
-    ),
-    OnboardingInfo(
-      title: "Personalized Designs",
-      description:
-          "You have complete control over your outfit's final look. Use our customization tools to adjust colors, patterns, and other details in your chosen design. Preview the results instantly and make adjustments until it's just right.",
-      image: "assets/images/image2.jpeg",
-    ),
-    OnboardingInfo(
-      title: "Seamless Experience",
-      description:
-          "Once you're satisfied with your design, you can place your order directly through the app. Select a trusted tailor or shop from our list of verified vendors to produce your garment.So Enjoy a seamless experience from concept to creation!",
-      image: "assets/images/image13.jpg",
-    ),
-  ];
-}
