--- Onbording/onbording_info.dart
@@ -1,11 +1,0 @@
-class OnboardingInfo {
-  final String title;
-  final String description;
-  final String image;
-
-  OnboardingInfo({
-    required this.title,
-    required this.description,
-    required this.image,
-  });
-}
