--- Onbording/onbording_view.dart
@@ -1,155 +1,0 @@
-import 'package:clothing_app/Onbording/onboarding_items.dart';
-import 'package:flutter/material.dart';
-
-class OnboardingView extends StatefulWidget {
-  const OnboardingView({super.key});
-
-  @override
-  State<OnboardingView> createState() => _OnboardingViewState();
-}
-
-class _OnboardingViewState extends State<OnboardingView> {
-  final controller = OnboardingItems();
-  final pageController = PageController();
-  int currentPage = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Stack(
-        children: [
-          Column(
-            children: [
-              Expanded(
-                child: PageView.builder(
-                  controller: pageController,
-                  itemCount: controller.items.length,
-                  onPageChanged: (index) {
-                    setState(() {
-                      currentPage = index;
-                    });
-                  },
-                  itemBuilder: (context, index) {
-                    final item = controller.items[index];
-                    return Column(
-                      children: [
-                        Flexible(
-                          flex: 3,
-                          child: FractionallySizedBox(
-                            widthFactor: 1.0,
-                            child: Image.asset(
-                              item.image,
-                              fit: BoxFit.cover,
-                            ),
-                          ),
-                        ),
-                        const SizedBox(height: 20),
-                        // Title
-                        Padding(
-                          padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                          child: Text(
-                            item.title,
-                            style: const TextStyle(
-                              fontSize: 24.0,
-                              fontWeight: FontWeight.bold,
-                              color: Colors.black, // Title color
-                            ),
-                            textAlign: TextAlign.center,
-                          ),
-                        ),
-                        const SizedBox(height: 20),
-                        // Description
-                        Flexible(
-                          flex: 1,
-                          child: Padding(
-                            padding:
-                                const EdgeInsets.symmetric(horizontal: 16.0),
-                            child: SingleChildScrollView(
-                              child: Text(
-                                item.description,
-                                style: Theme.of(context)
-                                    .textTheme
-                                    .bodyMedium
-                                    ?.copyWith(fontSize: 16.0),
-                                textAlign: TextAlign.center,
-                              ),
-                            ),
-                          ),
-                        ),
-                      ],
-                    );
-                  },
-                ),
-              ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: List.generate(
-                  controller.items.length,
-                  (index) => AnimatedContainer(
-                    duration: const Duration(milliseconds: 300),
-                    margin: const EdgeInsets.symmetric(horizontal: 4.0),
-                    height: 8.0,
-                    width: currentPage == index ? 16.0 : 8.0,
-                    decoration: BoxDecoration(
-                      color: currentPage == index ? Colors.black : Colors.grey,
-                      borderRadius: BorderRadius.circular(4.0),
-                    ),
-                  ),
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(vertical: 16.0),
-                child: SizedBox(
-                  width: double.infinity,
-                  child: Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                    child: ElevatedButton(
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Colors.black,
-                        foregroundColor: Colors.white,
-                      ),
-                      onPressed: () {
-                        if (currentPage < controller.items.length - 1) {
-                          pageController.nextPage(
-                            duration: const Duration(milliseconds: 300),
-                            curve: Curves.easeInOut,
-                          );
-                        } else {
-                          // Navigate to WelcomeScreen instead of LoginPage
-                          Navigator.pushReplacementNamed(context, '/welcome');
-                        }
-                      },
-                      child: Text(
-                        currentPage < controller.items.length - 1
-                            ? "Next"
-                            : "Finish",
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-          // "Skip" button
-          Positioned(
-            top: 40.0,
-            right: 16.0,
-            child: TextButton(
-              onPressed: () {
-                pageController.jumpToPage(controller.items.length - 1);
-              },
-              child: const Text(
-                "Skip",
-                style: TextStyle(
-                  fontSize: 16.0,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.white70,
-                ),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
