+++ clothing_app/lib/main.dart
@@ -1,9 +1,13 @@
-import 'package:clothing_app/Onbording/onbording_view.dart' hide HomePage;
+import 'package:clothing_app/Onbording/onbording_view.dart';
 import 'package:clothing_app/pages/singup.dart';
 import 'package:clothing_app/pages/login.dart';
 import 'package:clothing_app/pages/home.dart';
 import 'package:clothing_app/pages/customizing.dart'; // Import CustomizePage
-import 'package:clothing_app/pages/suggestions.dart'; // Import SuggestionPage
+import 'package:clothing_app/pages/store.dart';
+import 'package:clothing_app/pages/suggestions.dart';
+// ignore: duplicate_import
+import 'package:clothing_app/pages/store.dart';
+import 'package:clothing_app/pages/getstart.dart'; // Import SuggestionPage
 import 'package:flutter/material.dart';
 import 'dart:async';
 
@@ -37,6 +41,8 @@ class MyApp extends StatelessWidget {
         '/home': (context) => HomePage(),
         '/customize': (context) => const CustomizePage(),
         '/suggestion': (context) => const SuggestionPage(),
+        '/welcome': (context) => const WelcomeScreen(),
+        '/store': (context) => const StorePage(),
       },
       onUnknownRoute: (settings) {
         return MaterialPageRoute(
