+++ clothing_app/lib/pages/getstart.dart
