+++ clothing_app/lib/pages/suggestions.dart
@@ -1,5 +1,5 @@
+import 'package:clothing_app/pages/store.dart';
 import 'package:flutter/material.dart';
-import 'store.dart'; // Import the store.dart file
 
 class SuggestionPage extends StatelessWidget {
   const SuggestionPage({super.key});
@@ -7,225 +7,310 @@ class SuggestionPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        backgroundColor: Colors.white,
-        elevation: 0,
-        leading: IconButton(
-          icon: const Icon(Icons.menu, color: Colors.black),
-          onPressed: () {
-            // Handle menu button press
-          },
-        ),
-        title: const Text(
-          'DressMe',
-          style: TextStyle(
-            fontFamily: 'Cursive',
-            fontSize: 24,
-            color: Colors.black,
-            fontWeight: FontWeight.bold,
-          ),
-        ),
-        centerTitle: true,
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.search, color: Colors.black),
-            onPressed: () {
-              // Handle search button press
-            },
-          ),
-        ],
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          children: [
-            const SizedBox(height: 16),
-            Container(
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(16),
-                boxShadow: [
-                  BoxShadow(
-                    color: Colors.black.withOpacity(0.2),
-                    blurRadius: 8,
-                    offset: const Offset(2, 2),
-                  ),
-                ],
-              ),
-              child: Stack(
+      body: Stack(
+        children: [
+          Column(
+            children: [
+              // Header section
+              Stack(
                 children: [
-                  Transform.translate(
-                    offset: const Offset(0, -1),
-                    child: ClipRRect(
-                      borderRadius: BorderRadius.circular(20),
-                      child: Image.asset(
-                        'assets/images/image16.jpg',
-                        fit: BoxFit.cover,
-                        alignment: const Alignment(
-                            0, -0.1), // Adjust alignment to lower the imag
-                        width: double.infinity,
-                        height: 200,
+                  Container(
+                    color: const Color.fromARGB(255, 202, 220, 240),
+                    child: Padding(
+                      padding: const EdgeInsets.only(top: 10),
+                      child: ClipRRect(
+                        child: Image.asset(
+                          'assets/images/image27.png',
+                          fit: BoxFit.cover,
+                          height: 400,
+                          width: double.infinity,
+                          alignment: Alignment.centerLeft,
+                        ),
                       ),
                     ),
                   ),
                   Positioned(
-                    bottom: 0,
-                    left: 0,
-                    right: 0,
-                    child: Container(
-                      decoration: BoxDecoration(
-                        color: Colors.black.withOpacity(0.6),
-                        borderRadius: const BorderRadius.only(
-                          bottomLeft: Radius.circular(16),
-                          bottomRight: Radius.circular(16),
-                        ),
-                      ),
-                      padding: const EdgeInsets.all(12.0),
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          const Text(
-                            'PICK YOUR\nFAVOURITE ITEM ',
-                            style: TextStyle(
-                              color: Colors.white,
-                              fontSize: 25,
-                              fontWeight: FontWeight.bold,
-                            ),
+                    top: 200,
+                    left: 30,
+                    child: Column(
+                      children: const [
+                        Text(
+                          "P I C K U P\n Y O U R 'S ",
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontSize: 50,
+                            fontWeight: FontWeight.bold,
+                            shadows: [
+                              Shadow(
+                                color: Colors.black, // Shadow color
+                                offset: Offset(1, 1), // Shadow position (x, y)
+                                blurRadius: 5, // Shadow blur radius
+                              ),
+                            ],
                           ),
-                        ],
-                      ),
+                        ),
+                      ],
                     ),
                   ),
                 ],
               ),
-            ),
-            const SizedBox(height: 24),
-            Expanded(
-              child: GridView.builder(
-                gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-                  crossAxisCount: 2,
-                  crossAxisSpacing: 12,
-                  mainAxisSpacing: 12,
-                  childAspectRatio: 3 / 4,
+              const SizedBox(height: 16),
+              // "You also may like" section
+              Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 20),
+                child: Material(
+                  shape: const RoundedRectangleBorder(
+                    borderRadius:
+                        BorderRadius.vertical(top: Radius.circular(20)),
+                  ),
                 ),
-                itemCount: 3, // Number of product cards
-                itemBuilder: (context, index) {
-                  return ProductCard(index: index);
-                },
               ),
+              const SizedBox(height: 8),
+              // Product list with bottom space
+              Expanded(
+                child: ListView(
+                  scrollDirection: Axis.horizontal,
+                  children: const [
+                    ProductCard(
+                        image: 'assets/images/image30.png',
+                        title: "Puff long frock",
+                        price: "RS.4500.00"),
+                    ProductCard(
+                        image: 'assets/images/image33.png',
+                        title: "Cocoon Red Frock",
+                        price: "RS.6000.00"),
+                    ProductCard(
+                        image: 'assets/images/image32.png',
+                        title: "Wrap Dress",
+                        price: "RS. 6500.00"),
+                  ],
+                ),
+              ),
+              const SizedBox(height: 20), // Extra space below product list
+            ],
+          ),
+          // Black 3-line menu icon at the top left corner
+          Positioned(
+            top: 40,
+            left: 16,
+            child: IconButton(
+              icon: const Icon(Icons.menu, color: Colors.black),
+              onPressed: () {
+                // Handle menu button press
+                print("Menu button pressed");
+              },
             ),
-          ],
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+// Product Card with navigation to ProductDetailPage
+class ProductCard extends StatelessWidget {
+  final String image;
+  final String title;
+  final String price;
+
+  const ProductCard({
+    super.key,
+    required this.image,
+    required this.title,
+    required this.price,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: () {
+        Navigator.push(
+          context,
+          MaterialPageRoute(
+            builder: (context) => ProductDetailPage(
+              image: image,
+              title: title,
+            ),
+          ),
+        );
+      },
+      child: Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 8),
+        child: Container(
+          width: 140,
+          height: 140,
+          decoration: BoxDecoration(
+            color: Colors.white,
+            borderRadius: BorderRadius.circular(16),
+            boxShadow: [
+              BoxShadow(
+                color: Colors.black.withOpacity(0.1),
+                blurRadius: 8,
+                offset: const Offset(2, 2),
+              ),
+            ],
+            border: Border.all(color: Colors.black, width: 2),
+          ),
+          child: Column(
+            children: [
+              ClipRRect(
+                borderRadius: const BorderRadius.only(
+                  topLeft: Radius.circular(16),
+                  topRight: Radius.circular(16),
+                ),
+                child: Image.asset(
+                  image,
+                  fit: BoxFit.cover,
+                  height: 200,
+                  width: 140,
+                ),
+              ),
+              Expanded(
+                child: Container(
+                  padding: const EdgeInsets.all(8),
+                  color: Colors.white,
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Text(
+                        title,
+                        style: const TextStyle(
+                          fontSize: 14,
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                      const SizedBox(height: 4),
+                      Text(
+                        price,
+                        style: const TextStyle(
+                          fontSize: 12,
+                          color: Colors.grey,
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+            ],
+          ),
         ),
       ),
     );
   }
 }
 
-class ProductCard extends StatelessWidget {
-  final int index;
+class ProductDetailPage extends StatelessWidget {
+  final String image;
+  final String title;
 
-  const ProductCard({super.key, required this.index});
+  const ProductDetailPage({
+    super.key,
+    required this.image,
+    required this.title,
+  });
 
   @override
   Widget build(BuildContext context) {
-    final List<String> images = [
-      'assets/images/Sketch5.png', // Image for the first card
-      'assets/images/Sketch1.png', // Image for the second card
-      'assets/images/Sketch4.png', // Image for the third card
-    ];
-
-    return GestureDetector(
-      onTap: () {
-        showDialog(
-          context: context,
-          builder: (BuildContext context) {
-            return AlertDialog(
-              shape: RoundedRectangleBorder(
+    return Scaffold(
+      body: Stack(
+        children: [
+          // Fullscreen product image
+          Image.asset(
+            'assets/images/image15.jpg',
+            fit: BoxFit.cover,
+            height: double.infinity,
+            width: double.infinity,
+          ),
+          // Product Details Card
+          Positioned(
+            bottom: 20,
+            left: 16,
+            right: 16,
+            child: Container(
+              height: 220,
+              decoration: BoxDecoration(
+                color: Colors.white.withOpacity(0.9),
                 borderRadius: BorderRadius.circular(16),
               ),
-              contentPadding: const EdgeInsets.all(16),
-              content: Column(
-                mainAxisSize: MainAxisSize.min,
+              padding: const EdgeInsets.all(16),
+              child: Row(
+                crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
-                  // Smaller image of the product
+                  // Small preview image
                   ClipRRect(
-                    borderRadius: BorderRadius.circular(16),
+                    borderRadius: BorderRadius.circular(8),
                     child: Image.asset(
-                      images[index],
+                      image,
+                      height: 150,
+                      width: 100,
                       fit: BoxFit.cover,
-                      width: 150,
-                      height: 220,
                     ),
                   ),
-                  const SizedBox(height: 16),
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                    children: [
-                      // Customize text
-                      Column(
-                        children: const [
+                  const SizedBox(width: 16),
+                  // Product Title and Description
+                  Expanded(
+                    child: Padding(
+                      padding:
+                          const EdgeInsets.only(top: 20), // Adjusted padding
+                      child: Column(
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: [
                           Text(
-                            'Customize',
-                            style: TextStyle(
-                              color: Color.fromARGB(255, 70, 70, 70),
-                              fontSize: 16,
+                            title.toUpperCase(),
+                            style: const TextStyle(
+                              fontSize: 18,
                               fontWeight: FontWeight.bold,
                             ),
                           ),
+                          const SizedBox(height: 20), // Increased spacing
+                          const Text(
+                            "One button neck along long-sleeved waist female stitching dress",
+                            style: TextStyle(fontSize: 13),
+                            maxLines: 3,
+                            overflow: TextOverflow.ellipsis,
+                          ),
                         ],
                       ),
-                      // Next text with navigation
-                      GestureDetector(
-                        onTap: () {
-                          Navigator.pop(context); // Close the dialog
-                          Navigator.push(
-                            context,
-                            MaterialPageRoute(
-                              builder: (context) => const StorePage(),
-                            ),
-                          );
-                        },
-                        child: Column(
-                          children: const [
-                            Text(
-                              'Next',
-                              style: TextStyle(
-                                color: Color.fromARGB(255, 70, 70, 70),
-                                fontSize: 16,
-                                fontWeight: FontWeight.bold,
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ],
+                    ),
                   ),
                 ],
               ),
-            );
-          },
-        );
-      },
-      child: Container(
-        decoration: BoxDecoration(
-          color: Colors.white,
-          borderRadius: BorderRadius.circular(16),
-          boxShadow: [
-            BoxShadow(
-              color: Colors.black.withOpacity(0.1),
-              blurRadius: 8,
-              offset: const Offset(2, 2),
             ),
-          ],
-          border: Border.all(color: Colors.black, width: 1),
-        ),
-        child: ClipRRect(
-          borderRadius: BorderRadius.circular(16),
-          child: Image.asset(
-            images[index],
-            fit: BoxFit.cover,
           ),
-        ),
+          // Place Order Button
+          Positioned(
+            bottom: 40,
+            right: 42,
+            child: GestureDetector(
+              onTap: () {
+                // Navigate to StorePage
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => const StorePage(),
+                  ),
+                );
+              },
+              child: const Text(
+                "Place Order",
+                style: TextStyle(
+                  fontSize: 16,
+                  fontWeight: FontWeight.bold,
+                  color: Color.fromARGB(255, 0, 0, 0),
+                ),
+              ),
+            ),
+          ),
+          // Back Button
+          Positioned(
+            top: 40,
+            left: 16,
+            child: IconButton(
+              icon: const Icon(Icons.arrow_back, color: Colors.black),
+              onPressed: () => Navigator.pop(context),
+            ),
+          ),
+        ],
       ),
     );
   }
