+++ clothing_app/pubspec.yaml
@@ -15,6 +15,7 @@ dependencies:
   cupertino_icons: ^1.0.8
   flutter_native_splash: ^2.4.4
   flick_video_player: ^0.9.0
+  lottie: ^3.3.1
 
 dev_dependencies:
   flutter_test:
@@ -42,7 +43,16 @@ flutter:
     - assets/images/image15.jpg
     - assets/images/image16.jpg
     - assets/images/image17.jpg
-
+    - assets/images/image19.jpeg
+    - assets/images/image21.jpeg
+    - assets/images/image23.jpeg
+    - assets/images/image24.jpeg
+    - assets/images/image25.jpeg
+    - assets/images/image26.jpeg
+    - assets/images/image27.png
+    - assets/images/image30.png
+    - assets/images/image32.png
+    - assets/images/image33.png
     - assets/images/Sketch1.png
     - assets/images/Sketch2.png
     - assets/images/Sketch3.png
@@ -51,6 +61,17 @@ flutter:
     - assets/images/image11.png
 
     - assets/images/google_icon.png
+    - assets/animations/Animation2.json
+    - assets/animations/Animation1.json
+    - assets/animations/Animation6.json
+    - assets/animations/Animation7.json
+    - assets/animations/Animation8.json
+    - assets/animations/Animation4.json
+    - assets/animations/Animation5.json
+    - assets/animations/Animation9.json
+    - assets/animations/Animation10.json
+    - assets/animations/Animation11.json
+    - assets/animations/Animation15.json
 
   # Add font assets
   fonts:
