--- pages/singup.dart
@@ -1,172 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:lottie/lottie.dart';
-
-class SignUpPage extends StatefulWidget {
-  const SignUpPage({super.key});
-
-  @override
-  State<SignUpPage> createState() => _SignUpPageState();
-}
-
-class _SignUpPageState extends State<SignUpPage> {
-  final TextEditingController _emailController = TextEditingController();
-  final TextEditingController _passwordController = TextEditingController();
-  final TextEditingController _confirmPasswordController =
-      TextEditingController();
-  bool _isPasswordVisible = false;
-  bool _isConfirmPasswordVisible = false;
-
-  @override
-  void dispose() {
-    _emailController.dispose();
-    _passwordController.dispose();
-    _confirmPasswordController.dispose();
-    super.dispose();
-  }
-
-  void _handleSignUp() {
-    String email = _emailController.text.trim();
-    String password = _passwordController.text;
-    String confirmPassword = _confirmPasswordController.text;
-
-    if (email.isEmpty || password.isEmpty || confirmPassword.isEmpty) {
-      _showSnackBar("Please fill in all fields.");
-    } else if (!RegExp(r'^[^@]+@[^@]+\.[^@]+').hasMatch(email)) {
-      _showSnackBar("Please enter a valid email address.");
-    } else if (password.length < 6) {
-      _showSnackBar("Password must be at least 6 characters.");
-    } else if (password != confirmPassword) {
-      _showSnackBar("Passwords do not match. Please try again.");
-    } else {
-      Navigator.pushReplacementNamed(context, '/home');
-    }
-  }
-
-  void _showSnackBar(String message) {
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(content: Text(message)),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      body: SafeArea(
-        child: SingleChildScrollView(
-          padding: const EdgeInsets.symmetric(horizontal: 24.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: [
-              const SizedBox(height: 25),
-              Center(
-                child: Lottie.asset('assets/animations/Animation8.json',
-                    height: 200),
-              ),
-              const SizedBox(height: 20.0),
-              const Text(
-                'Register',
-                style: TextStyle(
-                    fontSize: 30,
-                    fontWeight: FontWeight.normal,
-                    color: Colors.black),
-                textAlign: TextAlign.left,
-              ),
-              const SizedBox(height: 5),
-              const Text(
-                'Please register to log in.',
-                style: TextStyle(fontSize: 16.0, color: Colors.grey),
-                textAlign: TextAlign.left,
-              ),
-              const SizedBox(height: 20),
-              _buildTextField(controller: _emailController, hint: "Email"),
-              const SizedBox(height: 16.0),
-              _buildPasswordField(
-                controller: _passwordController,
-                hint: "Password",
-                isVisible: _isPasswordVisible,
-                onVisibilityToggle: () {
-                  setState(() {
-                    _isPasswordVisible = !_isPasswordVisible;
-                  });
-                },
-              ),
-              const SizedBox(height: 16.0),
-              _buildPasswordField(
-                controller: _confirmPasswordController,
-                hint: "Confirm Password",
-                isVisible: _isConfirmPasswordVisible,
-                onVisibilityToggle: () {
-                  setState(() {
-                    _isConfirmPasswordVisible = !_isConfirmPasswordVisible;
-                  });
-                },
-              ),
-              const SizedBox(height: 24.0),
-              ElevatedButton(
-                onPressed: _handleSignUp,
-                style: ElevatedButton.styleFrom(
-                  backgroundColor: Colors.black,
-                  minimumSize: const Size(double.infinity, 45),
-                  shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(30)),
-                ),
-                child: const Text("Sign Up",
-                    style: TextStyle(fontSize: 16.0, color: Colors.white)),
-              ),
-              const SizedBox(height: 5.0),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  const Text("Have an account?",
-                      style: TextStyle(fontSize: 14.0, color: Colors.black)),
-                  TextButton(
-                    onPressed: () {
-                      Navigator.pushNamed(context, '/login');
-                    },
-                    child: const Text("Sign In",
-                        style: TextStyle(
-                            fontSize: 14.0,
-                            fontWeight: FontWeight.bold,
-                            color: Colors.black)),
-                  ),
-                ],
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildTextField(
-      {required TextEditingController controller, required String hint}) {
-    return TextField(
-      controller: controller,
-      decoration: InputDecoration(
-        hintText: hint,
-        border: OutlineInputBorder(borderRadius: BorderRadius.circular(12.0)),
-      ),
-    );
-  }
-
-  Widget _buildPasswordField({
-    required TextEditingController controller,
-    required String hint,
-    required bool isVisible,
-    required VoidCallback onVisibilityToggle,
-  }) {
-    return TextField(
-      controller: controller,
-      obscureText: !isVisible,
-      decoration: InputDecoration(
-        hintText: hint,
-        border: OutlineInputBorder(borderRadius: BorderRadius.circular(12.0)),
-        suffixIcon: IconButton(
-          icon: Icon(isVisible ? Icons.visibility : Icons.visibility_off),
-          onPressed: onVisibilityToggle,
-        ),
-      ),
-    );
-  }
-}
